support entires with no stats gen
